# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "CreationFailure": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "index": {
            "description": "Zero-based position in the submitted array; 0 for one object.",
            "title": "Index",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "index",
          "code",
          "message"
        ],
        "title": "CreationFailure",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MediaCreateResponse": {
        "properties": {
          "created": {
            "items": {
              "$ref": "#/components/schemas/MediaCreated"
            },
            "title": "Created",
            "type": "array"
          },
          "failed": {
            "items": {
              "$ref": "#/components/schemas/CreationFailure"
            },
            "title": "Failed",
            "type": "array"
          }
        },
        "title": "MediaCreateResponse",
        "type": "object"
      },
      "MediaCreated": {
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "renamed": {
            "default": false,
            "title": "Renamed",
            "type": "boolean"
          },
          "requested_filename": {
            "title": "Requested Filename",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "size",
          "index"
        ],
        "title": "MediaCreated",
        "type": "object"
      },
      "MediaDeletionResult": {
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "filename",
          "stats"
        ],
        "title": "MediaDeletionResult",
        "type": "object"
      },
      "MediaFile": {
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "mtime": {
            "title": "Mtime",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "size",
          "mtime"
        ],
        "title": "MediaFile",
        "type": "object"
      },
      "MediaList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MediaFile"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "title": "Next Cursor",
            "type": "string"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          }
        },
        "required": [
          "items",
          "stats"
        ],
        "title": "MediaList",
        "type": "object"
      },
      "MediaUpload": {
        "properties": {
          "data": {
            "title": "Data",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "title": "MediaUpload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "Required when an API key is configured in Tsunagi settings.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "description": "Alternative to the X-API-Key header.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "license": {
      "name": "MIT"
    },
    "title": "Tsunagi"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/media": {
      "get": {
        "description": "Media is a flat file namespace, not a queryable resource - filter with prefix/suffix rather than the select/where DSL.",
        "operationId": "listMedia",
        "parameters": [
          {
            "description": "Only names starting with this",
            "in": "query",
            "name": "prefix",
            "required": false,
            "schema": {
              "description": "Only names starting with this",
              "title": "Prefix",
              "type": "string"
            }
          },
          {
            "description": "Only names ending with this (e.g. '.mp3')",
            "in": "query",
            "name": "suffix",
            "required": false,
            "schema": {
              "description": "Only names ending with this (e.g. '.mp3')",
              "title": "Suffix",
              "type": "string"
            }
          },
          {
            "description": "Maximum results in this response. Omit to return all matches; no fixed upper cap.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "description": "Maximum results in this response. Omit to return all matches; no fixed upper cap.",
              "minimum": 1.0,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque next_cursor from the previous response. Omit to start at page one; malformed or empty cursors return 400.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "description": "Opaque next_cursor from the previous response. Omit to start at page one; malformed or empty cursors return 400.",
              "title": "Cursor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "List media files",
        "tags": [
          "Media"
        ],
        "x-permission": "read:media"
      },
      "post": {
        "description": "Accepts one upload object or an array. Each input provides exactly one of base64 'data', a server-local 'path' (disabled by default), or a 'url'. Always returns created and failed arrays with zero-based input indexes. Successful files stay stored when another is rejected. Anki renames on collision; use the returned filename. The configured upload-size limit applies per file. Media storage is not undoable.",
        "operationId": "storeMedia",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "items": {
                      "$ref": "#/components/schemas/MediaUpload"
                    },
                    "type": "array"
                  },
                  {
                    "$ref": "#/components/schemas/MediaUpload"
                  }
                ],
                "description": "One upload or an array of uploads",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Store one or more media files",
        "tags": [
          "Media"
        ],
        "x-permission": "write:media"
      }
    },
    "/v1/media/{filename}": {
      "delete": {
        "description": "Moves the file to Anki's media trash.",
        "operationId": "deleteMediaFile",
        "parameters": [
          {
            "description": "Media filename",
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "description": "Media filename",
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaDeletionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete a media file",
        "tags": [
          "Media"
        ],
        "x-permission": "write:media"
      },
      "get": {
        "description": "Streams the raw bytes with a guessed Content-Type. Note: when an API key is configured this URL cannot be used directly in <img src> - browsers can't attach the header; fetch() it and use createObjectURL.",
        "operationId": "getMediaFile",
        "parameters": [
          {
            "description": "Media filename",
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "description": "Media filename",
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Download a media file",
        "tags": [
          "Media"
        ],
        "x-permission": "read:media"
      }
    },
...
  },
...
}